Stable

+++ src/modules/camera_feedback/camera_feedback.hpp
@@ -45,9 +45,8 @@
 #include <stdbool.h>
 #include <poll.h>
 #include <mathlib/mathlib.h>
-#include <systemlib/systemlib.h>
 #include <systemlib/err.h>
-#include <systemlib/param/param.h>
+#include <parameters/param.h>
 
 #include <px4_config.h>
 #include <px4_defines.h>
@@ -62,7 +61,7 @@
 #include <uORB/topics/vehicle_local_position.h>
 #include <uORB/topics/vehicle_global_position.h>
 
-typedef enum : uint8_t {
+typedef enum : int32_t {
 	CAMERA_FEEDBACK_MODE_NONE = 0,
 	CAMERA_FEEDBACK_MODE_TRIGGER,
 	CAMERA_FEEDBACK_MODE_PWM
